Accessing empty square doesn't result in PHP Fatal anymore
